+++ CHANGES.rst
@@ -1,6 +1,15 @@
 Changelog
 =========
 
+0.2.0
+-----
+
+Please note: JSON files saved with the previous version are not compatible with this one; blocklists must be downloaded and saved again to work.
+
+- Keep track of source blocklist for each entry.
+- Add `bl_ids` and `bl_names` macros to the `lines` formatter.
+- Add a comment containing the source blocklist to each Mikrotik RouterOS address-list entry.
+
 0.1.0
 -----